Refactor report.rb to use one module per source file
