+++ Steepfile
@@ -4,6 +4,7 @@ target :lib do
   signature "sig"
 
   check "lib/lrama/bitmap.rb"
-  check "lib/lrama/report.rb"
+  check "lib/lrama/report/duration.rb"
+  check "lib/lrama/report/profile.rb"
   check "lib/lrama/warning.rb"
 end
