+++ sig/lrama/report/profile.rbs
@@ -1,0 +1,7 @@
+module Lrama
+  class Report
+    module Profile
+      def self.report_profile: { -> void } -> StackProf::result
+    end
+  end
+end
